01_HW_Firmware_Control/d011em-protocol-1-AksarbenCryunad/requirements.txt
pytest
gpiozero
pyserial
